// apps/frontend/src/api/generated/zod/model/pedcardForm.zod.ts
/**
 * Generated by orval from the Entropia Manager OpenAPI spec.
 * Do not edit manually.
 * Entropia Manager API
 */
import * as zod from 'zod';
import { PedcardType } from './pedcardType.zod.ts';

export const pedcardFormValueOneMin = 0;

export const pedcardFormValueTwoRegExp = new RegExp('^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$');


export const PedcardForm = zod.object({
  "type": PedcardType,
  "value": zod.union([zod.number().min(pedcardFormValueOneMin),zod.string().regex(pedcardFormValueTwoRegExp)]),
  "transactionId": zod.union([zod.string(),zod.null()]).optional()
});

export type PedcardForm = zod.input<typeof PedcardForm>;
export type PedcardFormOutput = zod.output<typeof PedcardForm>;